fix(fhir): TAMOC-382: Parallelize FHIR queue processing - Introduce FhirJobRunner to manage job runs individually rather than running as a batch
